Finish first stage of library project
This commit completes everything from the basic html to adding some
extra reactive features using JS. All of the main components are
finished.

The main issue right now is that the books aren't being stored in an
array which would make things like deleting a book easier because you
would already have the selector to it on hand. I skirted around this by
making each button and container have matching identifiers. This could
have been completed much easier by using the array method mentioned
earlier.

The other thing that needs to be completed going forward is that the
color pallette is absolutely ugly right now. I have chosen not to work
with it more than I already have. If this project is ever to be shown to
anyone, the styles need to be completely overhauled and the text font
changed.